% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_db.R
\name{can_i_write_table}
\alias{can_i_write_table}
\title{Write data to the database. But ask before!}
\usage{
can_i_write_table(schema)
}
\arguments{
\item{schema}{a schema in DB}
}
\description{
Do not work with spatial data! 
Use connect_to_db() default hence write to data instance
}
\examples{

\dontrun{
 # it is a function factory so first step is creating the function
 write_staging <- can_i_write_table("staging")
 # then using it
 write_staging("mtcars", mtcars, overwrite = TRUE)
}
}
