% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_dimensions.R
\name{compare_dimensions}
\alias{compare_dimensions}
\title{Compare dimensions of tables on the database}
\usage{
compare_dimensions(table_names, con)
}
\arguments{
\item{table_names}{A character vector of table names}

\item{con}{A valid database connection}
}
\value{
A data frame (data.table) containing the table name, count of rows, and count of columns
}
\description{
Compare dimensions of tables on the database
}
