% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_db.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Connect to CORI/RISI Postgres DB}
\usage{
connect_to_db(
  schema,
  host = "cori-risi-ad-postgresql.c6zaibvi9wyg.us-east-1.rds.amazonaws.com",
  dbname = "data",
  port = 5432
)
}
\arguments{
\item{schema}{A vector of schema names to add to the search path (e.g. 'sch_layer', sch_broadband', etc.)}

\item{host}{Database host, defaults to the host for the main MDA Potgres database.}

\item{dbname}{Database name to connect to. Defaults to main MDA Postgres database}

\item{port}{Port name to connect to. Defaults to port for main MDA Postgres database}
}
\value{
A pqconnection object
}
\description{
Connect to CORI/RISI Postgres DB
}
