% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_table_across_databases.R
\name{copy_table_across_databases}
\alias{copy_table_across_databases}
\title{Copy a table between two CORI/RISI Postgres DB instances}
\usage{
copy_table_across_databases(
  schema_name,
  table_name,
  from_db,
  to_db,
  grant_select_roles = c("read_only_access")
)
}
\arguments{
\item{schema_name}{schema name in from_db with target table that will be copied to same schema name in to_db}

\item{table_name}{name of the table that will be copied}

\item{from_db}{database name to copy the target table to}

\item{to_db}{database name to copy the target table to}

\item{grant_select_roles}{vector of role names that will be granted permission to select table rows (e.g. c("read_only_access","r_team"))}
}
\value{
a data.frame of the copied table
}
\description{
Copy a table between two CORI/RISI Postgres DB instances
}
