% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata_tables.R
\name{create_metadata_tables}
\alias{create_metadata_tables}
\title{Create template field and table metadata documents}
\usage{
create_metadata_tables(data_tbl = NULL, output_csv = TRUE)
}
\arguments{
\item{data_tbl}{Optionally provide a data frame, which will populate the \code{column_name} column with the column names}

\item{output_csv}{If \code{TRUE}, write field and table metadata templates to .CSV files in the working directory}
}
\value{
A named list with field and table metadata templates
}
\description{
The templates produced by this function match the format of the \code{metadata} schema
metadata tables. For new data sets, create templates, fill them out, then call either \code{write_db} providing
the completed templates as data frames or call \code{update_metadata} directly.
}
\examples{

\dontrun{
create_metadata_tables(mtcars, output_csv = FALSE)
}

}
