% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_on_db.R
\name{execute_on_postgres}
\alias{execute_on_postgres}
\title{Send query to postgres for execution. Vectorized over 'query'}
\usage{
execute_on_postgres(con, query)
}
\arguments{
\item{con}{A database connection}

\item{query}{A query with no expected return value (e.g. create, drop, alter)}
}
\description{
Send query to postgres for execution. Vectorized over 'query'
}
