% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_db_column_names.R
\name{get_db_column_names}
\alias{get_db_column_names}
\title{Get column names for a table from the information schema}
\usage{
get_db_column_names(con, table_name)
}
\arguments{
\item{con}{A valid database connection}

\item{table_name}{The name of the table to get column names for}
}
\value{
A data.frame of search results, including schema name, table name, and column names
}
\description{
Get column names for a table from the information schema
}
