% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_dimensions.R
\name{get_dims}
\alias{get_dims}
\title{Get the dimensions of a table on the database}
\usage{
get_dims(table_name, con)
}
\arguments{
\item{table_name}{The name of table on the database}

\item{con}{A database connection}
}
\value{
A one row data frame with the table name, a row count, and a column count
}
\description{
Get the dimensions of a table on the database
}
