% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sql.R
\name{get_sql}
\alias{get_sql}
\title{Originally from:
  https://stackoverflow.com/questions/44853322/how-to-read-the-contents-of-an-sql-file-into-an-r-script-to-run-a-query#answer-44886192}
\usage{
get_sql(filepath)
}
\arguments{
\item{filepath}{path to a SQL script (.sql extension)}
}
\value{
A string vector of SQL statements extracted from the script
}
\description{
"I've had trouble with reading sql files myself, and have found that often times the syntax gets broken if there
  are any single line comments in the sql. Since in R you store the sql statement as a single line string, if there
  are any double dashes in the sql it will essentially comment out any code after the double dash.
}
\details{
This is a function that I typically use whenever I am reading in a .sql file to be used in R."
}
