% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_bucket.R
\name{list_bucket}
\alias{list_bucket}
\title{List every "keys" from a bucket}
\usage{
list_bucket(bucket_name)
}
\arguments{
\item{bucket_name}{string, example "test-coridata""}
}
\value{
a data frame with a key and last_modified columns
}
\description{
List every "keys" from a bucket
}
\examples{

\dontrun{
 s3_asset <- list_bucket("test-coridata")
}

}
