% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_session_metadata.R
\name{load_session_metadata}
\alias{load_session_metadata}
\title{Loads session metadata from temp directory}
\usage{
load_session_metadata()
}
\value{
Returns table, field, and source metadata in a named list.
}
\description{
Loads session metadata from temp directory
}
