% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_aliased_query.R
\name{make_aliased_query}
\alias{make_aliased_query}
\title{Create a quoted SQL query with column aliasing from named list}
\usage{
make_aliased_query(named_list, source_table, command = "select", con)
}
\arguments{
\item{named_list}{A named list}

\item{source_table}{The name of the table to select from}

\item{command}{What to do with the aliased query (e.g. 'create table foo as'). Defaults to 'select'.}

\item{con}{A database connection for `glue::glue_sql()`}
}
\value{
SQL query string
}
\description{
Create a quoted SQL query with column aliasing from named list
}
