% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_db.R
\name{prepare_to_write_table}
\alias{prepare_to_write_table}
\title{Write data to the database. But ask before!}
\usage{
prepare_to_write_table(write_table_function)
}
\arguments{
\item{schema}{a schema in DB}
}
\description{
Do not work with spatial data! 
Use connect_to_db() default hence write to data instance
}
\examples{

\dontrun{
 # This is a function factory that takes a callback function
 # and returns a wrapper function, so first step is defiining the callback
 # and passing it to generate the wrapper function:
 write_staging <- prepare_to_write_table(
     function(table_name, dat, ...) {
         con <- cori.db::connect_to_db("staging") # Define the db/schema connection here
         on.exit(DBI::dbDisconnect(con), add = TRUE)
         DBI::dbWriteTable(con, table_name, dat, ...)
     }
 )
 # ...then use it:
 write_staging("mtcars", mtcars, overwrite = TRUE)
 # ... which should prompt the user before running the callback function
}
}
