% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_s3_object.R
\name{put_s3_object}
\alias{put_s3_object}
\title{put an a local file into an s3 bucket}
\usage{
put_s3_object(bucket_name, object, key = object, ...)
}
\arguments{
\item{bucket_name}{string, a bucket name}

\item{object}{string, file (path) that you want to upload}

\item{key}{string, how will be named the key in s3 bucket, by default same as object}

\item{...}{other arguments from paws's put_object()}
}
\value{
return invisibly the response from AWS
}
\description{
put an a local file into an s3 bucket
}
\examples{

\dontrun{
 txt <- put_s3_object("test-coridata", "blabla.txt")
see what it returns
}

}
