% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_acs.R
\name{read_acs}
\alias{read_acs}
\title{Read MDA ACS data from the database}
\usage{
read_acs(con, variables, year = NULL, geography = "county", tidy = TRUE)
}
\arguments{
\item{con}{A database connection}

\item{variables}{A vector of variables to pull}

\item{year}{The desired year of data. If not provided, function pulls latest available year}

\item{geography}{Desired geography, currently one of county or place (as of 1/7/21)}

\item{tidy}{If TRUE return data as a tidy dataframe. Otherwise, retrieve unmodified normalized data}
}
\value{
A data frame
}
\description{
Read MDA ACS data from the database
}
