% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_db.R
\name{read_db}
\alias{read_db}
\title{Read a database table (convenience wrapper for DBI::dbReadTable and sf::st_read)}
\usage{
read_db(con, table, spatial = FALSE)
}
\arguments{
\item{con}{A database connection}

\item{table}{The name of the table to read}

\item{spatial}{If TRUE return a spatial data frame}
}
\value{
A data.table or sf class dataframe
}
\description{
Read a database table (convenience wrapper for DBI::dbReadTable and sf::st_read)
}
