% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_s3_object.R
\name{read_s3_object}
\alias{read_s3_object}
\title{read a text file (csv) in a s3 object in memory}
\usage{
read_s3_object(bucket_name, key, ...)
}
\arguments{
\item{bucket_name}{string, a bucket name}

\item{key}{string, object/file that you want to download}

\item{...}{extra argument to pass on read.csv()}
}
\value{
return a data frame
}
\description{
read a text file (csv) in a s3 object in memory
}
\details{
works only for content type
}
\examples{

\dontrun{
 my_csv <- read_s3_object("test-coridata", "data-1715776270877.csv")
}

}
