% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sql.R
\name{read_sql}
\alias{read_sql}
\title{Read SQL file}
\usage{
read_sql(filepaths)
}
\arguments{
\item{filepaths}{Path to .SQL or .TXT file to read. Can be a vector or list of files}
}
\value{
A character vector of SQL queries
}
\description{
Read SQL file
}
