% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{schema_size_db}
\alias{schema_size_db}
\title{Listing DB schema size}
\usage{
schema_size_db(schema = "public", ...)
}
\arguments{
\item{schema}{string should be set to public (it is by default)}

\item{...}{argument to be passed to connect_to_db()}
}
\value{
a data frame with two colums schema_name and size of said schema
}
\description{
Listing DB schema size
}
\examples{

\dontrun{
  schema_size_db()
}

}
