% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_db_tables.R
\name{search_db_tables}
\alias{search_db_tables}
\title{Search for table names in Postgres information schema}
\usage{
search_db_tables(con, table_search_string = "\%", schema = NULL)
}
\arguments{
\item{con}{A valid database connection}

\item{table_search_string}{A search string, defaults to a wildcard matching all table names. Searches using SQL 'like' so use percent sign for partial matching}

\item{schema}{Optional name of schema to search in}
}
\value{
Data.frame of search results, including table and schema names
}
\description{
Search for table names in Postgres information schema
}
