% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_aws_credentials.R
\name{set_aws_credentials}
\alias{set_aws_credentials}
\title{Install aws credentials in your .Renviron file and
load credentials into the current environment.
This actions will overwrite any values currently
stored in AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY.}
\usage{
set_aws_credentials(keyID, accesKey, region = "us-east-1", install = TRUE)
}
\arguments{
\item{keyID}{A valid Key ID}

\item{accesKey}{A valid Access key}

\item{region}{A valid AWS region (default set )}

\item{install}{If TRUE, save credentials to .Renviron. Otherwise, set per session.}
}
\value{
keyID, accesKey and region, invisibly
}
\description{
Install aws credentials in your .Renviron file and
load credentials into the current environment.
This actions will overwrite any values currently
stored in AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY.
}
