% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_db_credentials.R
\name{set_db_credentials}
\alias{set_db_credentials}
\title{Install database credentials in your .Renviron file}
\usage{
set_db_credentials(username, password, install = TRUE)
}
\arguments{
\item{username}{A valid Postgres username}

\item{password}{A valid Postgres password}

\item{install}{If TRUE, save credentials to .Renviron. Otherwise, set per session.}
}
\value{
Username and password, invisibly
}
\description{
Install database credentials in your .Renviron file
}
