% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_db.R
\name{simple_write_to_db}
\alias{simple_write_to_db}
\title{Open a connection and write a table to specific schema}
\usage{
simple_write_to_db(
  schema,
  name,
  dta,
  overwrite = TRUE,
  DB_instance = "data",
  ...
)
}
\arguments{
\item{schema}{Name of the schema to write}

\item{name}{Name of the table to write}

\item{dta}{A data frame that will be wrote}

\item{overwrite}{`TRUE` by default, overwrite an existing table}

\item{DB_instance}{default `data`, passed to the connection to specify DB instance}

\item{...}{other arguments for DBI::dbWriteTable()}
}
\description{
A simpel wrapper function around open a connexion to the DB data, write a table and 
close the connexion.
}
