% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_join.R
\name{update_join}
\alias{update_join}
\title{Update a database table using an update join}
\usage{
update_join(
  old_tbl,
  new_tbl,
  variable_to_update,
  variable_to_update_from = NULL,
  key,
  con
)
}
\arguments{
\item{old_tbl}{Name of the table with the old values}

\item{new_tbl}{Name of the table with the new values}

\item{variable_to_update}{The variable to update in old_tbl}

\item{variable_to_update_from}{The variable in new_tbl to use to update the values in old_tbl}

\item{key}{Key variable to join on}

\item{con}{A Postgres connection}
}
\description{
Update a database table using an update join
}
