% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_metadata.R
\name{update_metadata}
\alias{update_metadata}
\title{Update metadata tables}
\usage{
update_metadata(con, field_meta, table_meta, source_meta = NULL)
}
\arguments{
\item{con}{A database connection}

\item{field_meta}{A data frame matching the structure for field metadata generated by \code{create_metadata_tables()}}

\item{table_meta}{A data frame matching the structure for table metadata generated by \code{create_metadata_tables()}}

\item{source_meta}{A data frame matching the structure for source metadata generated by \code{create_metadata_tables()}}
}
\description{
Update metadata tables
}
