% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_db.R
\name{write_db}
\alias{write_db}
\title{Write data to the database. Optionally update metadata tables in the process}
\usage{
write_db(
  con,
  name,
  dta,
  overwrite = FALSE,
  append = FALSE,
  spatial = FALSE,
  field_meta = NULL,
  table_meta = NULL,
  source_meta = NULL
)
}
\arguments{
\item{con}{A database connection}

\item{name}{Name of the table to write to}

\item{dta}{A data frame or sf data frame}

\item{overwrite}{If \code{TRUE}, overwrite an existing table}

\item{append}{If \code{TRUE}, append to an existing table}

\item{spatial}{If \code{TRUE}, write as spatial data}

\item{field_meta}{An optional data frame matching the structure for field metadata generated by \code{create_metadata_tables()}}

\item{table_meta}{An optional data frame matching the structure for table metadata generated by \code{create_metadata_tables()}}

\item{source_meta}{An optional data frame matching the structure for source metadata generated by \code{create_metadata_tables()}}
}
\description{
Write data to the database. Optionally update metadata tables in the process
}
