% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_s3_object.R
\name{write_s3_object}
\alias{write_s3_object}
\title{write a data.frame into a csv in a s3 bucket}
\usage{
write_s3_object(bucket_name, data_frame, key, ...)
}
\arguments{
\item{bucket_name}{string, a bucket name}

\item{data_frame}{an R object of class "data.frame", sf object are excluded}

\item{key}{string, object/file that you want to download}

\item{...}{extra argument to pass on write.csv()}
}
\value{
return invisibly the response from AWS
}
\description{
write a data.frame into a csv in a s3 bucket
}
\details{
it is using tempfile() shenanigans
}
\examples{

\dontrun{
 put_s3_object("test-coridata", cars, "cars.csv")
}

}
