% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_session_codebook.R
\name{write_session_codebook}
\alias{write_session_codebook}
\title{Writing session codebook to either an excel document, located in the working directory,
or googlesheet in your home drive.}
\usage{
write_session_codebook(name, out = "excel")
}
\arguments{
\item{name}{Name of the output file. No file extension required!}

\item{out}{One of "excel" or "g_drive". Defines output file type.
The default being a .xlsx, explicitly defined as "excel",
or the user can define a googlesheet with "g_drive". Output file will be in snake_case.}
}
\description{
Writing session codebook to either an excel document, located in the working directory,
or googlesheet in your home drive.
}
