% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_table_across_schemas.R
\name{copy_table_across_schemas}
\alias{copy_table_across_schemas}
\title{Copy a table from one schema to another within the same CORI/RISI Postgres DB instance
... will fail if table already exists in to_schema:
ERROR:  relation "..." already exists}
\usage{
copy_table_across_schemas(
  from_schema,
  to_schema,
  table_name,
  grant_select_roles = c("read_only_access"),
  overwrite = FALSE
)
}
\arguments{
\item{from_schema}{schema name to copy the target table from}

\item{to_schema}{schema name to copy the target table to}

\item{table_name}{name of the table that will be copied}

\item{grant_select_roles}{vector of role names that will be granted permission to select table rows (e.g. c("read_only_access","r_team"))}

\item{overwrite}{if TRUE and table exists in to_schema, it will be dropped and recreated (default = FALSE)}
}
\value{
the status of the copy transaction
}
\description{
Copy a table from one schema to another within the same CORI/RISI Postgres DB instance
... will fail if table already exists in to_schema:
ERROR:  relation "..." already exists
}
