% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_s3_object.R
\name{get_s3_object}
\alias{get_s3_object}
\title{get object hosted in s3 downloaded}
\usage{
get_s3_object(bucket_name, key, file_path, ...)
}
\arguments{
\item{bucket_name}{string, a bucket name}

\item{key}{string, object/file that you want to download}

\item{file_path}{string, a path where you want the file to be downloaded ("~/Documents/my_proj/"")}

\item{...}{other arguments from paws's downloadfile}
}
\value{
return invisibly the path where the file has been downloaded
}
\description{
get object hosted in s3 downloaded
}
\details{
If file_path is not provided it will the object in the curent directory
}
\examples{

\dontrun{
 txt <- get_s3_key("test-coridata", "blabla.txt")
 txt
}

}
