% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_bucket.R
\name{list_bucket}
\alias{list_bucket}
\title{List every "keys" from a bucket
 
load credentials into the current environment.
This actions will overwrite any values currently
stored in AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY.}
\usage{
list_bucket(bucket_name)
}
\arguments{
\item{Bucket}{string, example "test-coridata""}
}
\value{
a data frame with ...
}
\description{
List every "keys" from a bucket
 
load credentials into the current environment.
This actions will overwrite any values currently
stored in AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY.
}
\examples{

\dontrun{

}

}
