% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_s3_buckets.R
\name{list_s3_buckets}
\alias{list_s3_buckets}
\title{Listing s3 bucket CORI draft}
\usage{
list_s3_buckets()
}
\value{
a data frame with four columns: bucket_name, creation_date, group, owner (TODO: add lifecycle)
}
\description{
Listing s3 bucket CORI draft
}
