% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_db.R
\name{prepare_to_write_table}
\alias{prepare_to_write_table}
\title{Function factory to save you an argument (schema)}
\usage{
prepare_to_write_table(schema, DB_instance = "data", ...)
}
\arguments{
\item{schema}{Name of the schema}

\item{DB_instance}{default `data`, passed to the connection to specify DB instance}

\item{...}{other arguments for DBI::dbWriteTable()}
}
\description{
Function factory to save you an argument (schema)
}
\examples{

\dontrun{
# This is a function factory that takes a schema
# and returns a wrapper function, so first step is defiining
# the callback and passing it to generate the wrapper function:
write_staging <- prepare_to_write_table("staging")

# ...then use it:
write_staging("mtcars", mtcars, overwrite = TRUE)

# ... which should prompt the user before running the callback function to write data to db
}
}
