% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_s3_object.R
\name{put_s3_object}
\alias{put_s3_object}
\title{put a local file into an s3 bucket}
\usage{
put_s3_object(bucket_name, key, file_path, ...)
}
\arguments{
\item{bucket_name}{string, a bucket name}

\item{key}{string name of the key in s3 bucket}

\item{file_path}{string, file (path) that you want to upload}

\item{...}{other arguments from paws's put_object()}
}
\value{
return invisibly the response from AWS
}
\description{
put a local file into an s3 bucket
}
\examples{

\dontrun{
 txt <- put_s3_object("test-coridata", "blabla.txt" ,"blabla.txt")
}

}
