% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/put_s3_object.R
\name{put_s3_object}
\alias{put_s3_object}
\title{put a local file into an S3 bucket}
\usage{
put_s3_object(bucket_name, s3_key_path, file_path, ...)
}
\arguments{
\item{bucket_name}{string, S3 bucket name}

\item{s3_key_path}{string, intended path + name of the file within S3 bucket}

\item{file_path}{string, local path + name of the file that you want to upload}

\item{...}{other arguments to paws's put_object()}
}
\value{
return invisibly the response from AWS
}
\description{
put a local file into an S3 bucket
}
\examples{

\dontrun{
 txt <- put_s3_object("test-coridata", "blabla.txt" ,"blabla.txt")
}

}
