% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_db.R
\name{read_db}
\alias{read_db}
\title{Read a database table (convenience wrapper for DBI::dbReadTable and sf::st_read)}
\usage{
read_db(con, table = NULL, query = NULL, spatial = FALSE)
}
\arguments{
\item{con}{A database connection}

\item{table}{Optional name of a table to read}

\item{query}{An optional SQL query to execute against the database. Use to read subsets of large tables}

\item{spatial}{If TRUE return an SF spatial data frame}
}
\value{
A data.table or sf class dataframe
}
\description{
Read a database table (convenience wrapper for DBI::dbReadTable and sf::st_read)
}
