% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_join.R
\name{update_join}
\alias{update_join}
\title{Update a database table using an update join}
\usage{
update_join(
  table_to_update,
  table_to_update_from,
  variable_update_map,
  key,
  con
)
}
\arguments{
\item{table_to_update}{Name of the table with the old values}

\item{table_to_update_from}{Name of the table with the new values}

\item{variable_update_map}{A list of characters corresponding to shared names of columns to update, or a named list of key value pairs in the form \code{list(variable_name_to_update = 'variable_name_to_update_from')}}

\item{key}{Key variable to join on or named list in the form \code{list(old_table_key = 'new_table_key')}}

\item{con}{A Postgres connection}
}
\description{
Update a database table using an update join
}
